refactor: extract PanelHeader component to new file
