+++ src/views/components/index.ts
@@ -16,3 +16,4 @@ export {
 	ScreenshotsList
 } from '@views/components/ScreenshotsList/ScreenshotsList.tsx'
 export { PanelBanner } from '@views/components/PanelBanner/PanelBanner.tsx'
+export { PanelHeader } from '@views/components/PanelHeader/PanelHeader.tsx'
